Add StripePayoutable concern
